pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rt-mcmc-project"
version = "0.1.0"
description = "Bayesian Rt estimation with Metropolis-within-Gibbs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
